split out racing TestSignaturePass
